Move router token and interfaces into application tokens.
